Port t_lpback.sh test to tests repo
This test was historically used as a package test (the UUT was the package
build per configure specifications), but it lived in the tests subdir, and
refered to the $top_builddir/src/openvpn binary from within the test. This
caused a dependency issue. The package had to be built before make check
could be run.

The awkward referencing of the built binary from the test dir, and the
dependency issue were clues that logically this test was misfiled.

The t_lpback.sh checks the target binary succeeds when running
with --crypto <x> --test-crypto

The misfiling is that this test is both an integration test, and a package
test. It can be run against the package as configured (if crypto is enabled
in the configuration), and it can be run against relevant integration
targets like those in does-it-build, to check that all possible
integrations are well behaved.

It runs as a package test if --enable-unit-tests=yes or =all, and it runs
as integration test if --enable-unit-tests=all

This commit enables --test-crypto both against the configured and built
package (as a package test), and against selected integration test builds
(p2p, multi combo, multi client, with openssl and with mbedssl).

+++ tests/integration/does-it-build/t_lpback.sh
@@ -20,26 +20,49 @@
 # 02110-1301, USA.
 
 set -eu
-top_builddir="${top_builddir:-..}"
+
+[[ ${#@} == 1 ]] || {
+    echo "Usage: $0 <binary name>"
+    exit 99
+}
+bin=${1}
+
+echo -n "Check support for broken cipher mitigations ... "
+if ! "$bin" | grep -q "SSL+mitigations" ; then
+    echo "FAILED"
+    exit 1
+else
+    echo "PASS"
+fi
+
+if [[ ${IMPLEMENTED_tls_crypt_2_5:+isset} ]] ; then
 trap "rm -f key.$$ tc-server-key.$$ tc-client-key.$$ log.$$ ; trap 0 ; exit 77" 1 2 15
 trap "rm -f key.$$ tc-server-key.$$ tc-client-key.$$ log.$$ ; exit 1" 0 3
+else
+trap "rm -f key.$$ log.$$ ; trap 0 ; exit 77" 1 2 15
+trap "rm -f key.$$ log.$$ ; exit 1" 0 3
+fi
 
 # Get list of supported ciphers from openvpn --show-ciphers output
-CIPHERS=$(${top_builddir}/src/openvpn/openvpn --show-ciphers | \
+CIPHERS=$($bin --show-ciphers | \
             sed -e '/The following/,/^$/d' -e s'/ .*//' -e '/^[[:space:]]*$/d')
 
 # Also test cipher 'none'
 CIPHERS=${CIPHERS}$(printf "\nnone")
 
-"${top_builddir}/src/openvpn/openvpn" --genkey secret key.$$
+if [[ ${IMPLEMENTED_tls_crypt_2_5:+isset} ]] ; then
+"$bin" --genkey secret key.$$
+else
+"$bin" --genkey --secret key.$$
+fi
 set +e
 
 e=0
 for cipher in ${CIPHERS}
 do
     echo -n "Testing cipher ${cipher}... "
-    ( "${top_builddir}/src/openvpn/openvpn" --test-crypto --secret key.$$ --cipher ${cipher} ) >log.$$ 2>&1
-    if [ $? != 0 ] ; then
+    ( "$bin" --test-crypto --secret key.$$ --cipher ${cipher} ) >log.$$ 2>&1
+    if [ $? != 0 ] && ! grep -q "$cipher system implementation is faulty, and no mitigation is available" log.$$; then
         echo "FAILED"
         cat log.$$
         e=1
@@ -48,8 +71,10 @@ do
     fi
 done
 
+if [[ ${IMPLEMENTED_tls_crypt_2_5:+isset} ]] ; then
+
 echo -n "Testing tls-crypt-v2 server key generation..."
-"${top_builddir}/src/openvpn/openvpn" \
+"$bin" \
     --genkey tls-crypt-v2-server tc-server-key.$$ >log.$$ 2>&1
 if [ $? != 0 ] ; then
     echo "FAILED"
@@ -60,7 +85,7 @@ else
 fi
 
 echo -n "Testing tls-crypt-v2 key generation (no metadata)..."
-"${top_builddir}/src/openvpn/openvpn" --tls-crypt-v2 tc-server-key.$$ \
+"$bin" --tls-crypt-v2 tc-server-key.$$ \
     --genkey tls-crypt-v2-client tc-client-key.$$ >log.$$ 2>&1
 if [ $? != 0 ] ; then
     echo "FAILED"
@@ -78,7 +103,7 @@ while [ $i -lt 732 ]; do
     i=$(expr $i + 1)
 done
 echo -n "Testing tls-crypt-v2 key generation (max length metadata)..."
-"${top_builddir}/src/openvpn/openvpn" --tls-crypt-v2 tc-server-key.$$ \
+"$bin" --tls-crypt-v2 tc-server-key.$$ \
     --genkey tls-crypt-v2-client tc-client-key.$$ "${METADATA}" \
     >log.$$ 2>&1
 if [ $? != 0 ] ; then
@@ -90,5 +115,11 @@ else
 fi
 
 rm key.$$ tc-server-key.$$ tc-client-key.$$ log.$$
+
+else
+
+rm key.$$ log.$$
+
+fi
 trap 0
 exit $e
